+++ test/data/examples/script/dest/bin/flavoring
@@ -1,0 +1,2 @@
+#!/bin/bash
+printf 'Who likes %s?' 'licorice'
