+++ test/data/examples/script/src/bin/flavoring
@@ -1,0 +1,2 @@
+#!/bin/bash
+printf 'Who likes %s?' '{{ dotplate.vars.flavor }}'
